+++ polars/polars-lazy/src/physical_plan/expressions/default.rs
@@ -8,31 +8,6 @@ use rayon::prelude::*;
 use std::borrow::Cow;
 use std::sync::Arc;
 
-pub struct CastExpr {
-    pub(crate) input: Arc<dyn PhysicalExpr>,
-    pub(crate) data_type: DataType,
-}
-
-impl CastExpr {
-    pub fn new(input: Arc<dyn PhysicalExpr>, data_type: DataType) -> Self {
-        Self { input, data_type }
-    }
-}
-
-impl PhysicalExpr for CastExpr {
-    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
-        let series = self.input.evaluate(df, state)?;
-        series.cast_with_datatype(&self.data_type)
-    }
-    fn to_field(&self, input_schema: &Schema) -> Result<Field> {
-        self.input.to_field(input_schema)
-    }
-
-    fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation> {
-        Ok(self)
-    }
-}
-
 pub struct TernaryExpr {
     pub predicate: Arc<dyn PhysicalExpr>,
     pub truthy: Arc<dyn PhysicalExpr>,
